+++ secret-diary/stage1/src/test/java/org/hyperskill/secretdiary/Stage1UnitTest.kt
@@ -29,7 +29,6 @@ class Stage1UnitTest : AbstractUnitTest<MainActivity>(MainActivity::class.java) 
 
         val messageBtnSaveWrongText = "The text of btnSave should be \"Save\""
         assertEquals(messageBtnSaveWrongText, "Save", btnSave.text.toString())
-
         btnSave
     }
