+++ secret-diary/stage4/src/test/java/org/hyperskill/secretdiary/Stage4UnitTest.kt
@@ -7,6 +7,7 @@ import android.widget.EditText
 import android.widget.TextView
 import kotlinx.datetime.Clock
 import org.hyperskill.secretdiary.internals.AbstractUnitTest
+import org.hyperskill.secretdiary.internals.CustomClockSystemShadow
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
