+++ secret-diary/stage5/src/test/java/org/hyperskill/secretdiary/Stage5UnitTest.kt
@@ -4,6 +4,7 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.TextView
 import org.hyperskill.secretdiary.internals.AbstractUnitTest
+import org.hyperskill.secretdiary.internals.CustomClockSystemShadow
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
 import org.junit.Test
