+++ secret-diary/stage5/src/test/java/org/hyperskill/secretdiary/internals/CustomClockSystemShadow.kt
@@ -1,4 +1,4 @@
-package org.hyperskill.secretdiary
+package org.hyperskill.secretdiary.internals
 
 import android.os.SystemClock
 import kotlinx.datetime.Clock
