+++ secret-diary/stage5/task-info.yaml
@@ -28,5 +28,5 @@ files:
   visible: true
 - name: src/main/java/org/hyperskill/secretdiary/data/Diary.kt
   visible: true
-- name: src/test/java/org/hyperskill/secretdiary/CustomClockSystemShadow.kt
+- name: src/test/java/org/hyperskill/secretdiary/internals/CustomClockSystemShadow.kt
   visible: false
